feat: guest user support
# Prompts
- We now need to create guest account for anonymous users.

First we need to update Users table:  `google_id` => `oauth_id`, new field `oauth_type` (which has `google` and `guest` as potential values).

Now, for each anonymous user visiting the site, we create a guest account for them, which has `oauth_type` as `guest` and `oauth_id` as the user's session fingerprint based on IP and relevant browser info.

No need to be backward-compatbile, as current db will be purged.

- Don't consider a guest account as logged-in user. Still show the Oauth login link/button, prompting user to log in.